Back-end/requirements.txt
djangorestframework-simplejwt
djangorestframework 
requests
django-cors-headers
gunicorn
httpx